payment-backend/requirements.txt
# Web
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

# Kafka
confluent-kafka==2.4.0

# Redis
redis==5.0.4

# PostgreSQL
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation / serialisation
pydantic==2.7.1
pydantic-settings==2.3.1

# ML
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.2
pyarrow==23.0.1
scikit-learn==1.4.2

# Observability
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0

# Stripe
stripe==9.12.0

# Spark
pyspark==3.5.8

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-timeout==2.3.1
httpx==0.27.0

# dbt
dbt-core==1.8.9
dbt-postgres==1.8.2
# dbt-bigquery==1.8.3 — add in Phase 11 for prod profile
requests>=2.31.0
